+++ src/features/acf/frontend/common/WithFilters/model.ts
@@ -1,8 +1,7 @@
 //TODO: Pensar melhor nesse import, tá muito nested.
-import type * as diabetes from "@/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable";
+import type { AppliedFiltersCoaps } from "@/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/CoapsDataTable";
+import type { AppliedFiltersCoeq } from "@/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/CoeqDataTable";
 
 /// Contém todos os possíveis conjuntos de filtros.
 /// Eventualmente, este Union type deve ter um item para cada combinação indicador x perfil
-export type AppliedFilters =
-    | diabetes.AppliedFiltersCoeq
-    | diabetes.AppliedFiltersCoaps;
+export type AppliedFilters = AppliedFiltersCoeq | AppliedFiltersCoaps;
