--- src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/index.tsx
@@ -1,4 +1,0 @@
-export { CoapsDataTable } from "./modules/CoapsDataTable";
-export type { AppliedFiltersCoaps } from "./modules/CoapsDataTable";
-export { CoeqDataTable } from "./modules/CoeqDataTable";
-export type { AppliedFiltersCoeq } from "./modules/CoeqDataTable";
