+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/common/CpfOrBirthdayFormatter/__tests__/CpfOrBirthdayFormatter.test.tsx
